+++ livechat/configuration/api/v35.py
@@ -1545,26 +1545,6 @@ class ConfigurationApiV35(HttpClient):
                                  json=payload,
                                  headers=headers)
 
-    def get_company_details(self,
-                            payload: dict = None,
-                            headers: dict = None) -> httpx.Response:
-        ''' Gets company details of the license.
-            Args:
-                payload (dict): Custom payload to be used as request's data.
-                                It overrides all other parameters provided for the method.
-                headers (dict): Custom headers to be used with session headers.
-                                They will be merged with session-level values that are set,
-                                however, these method-level parameters will not be persisted across requests.
-            Returns:
-                httpx.Response: The Response object from `httpx` library,
-                                which contains a server's response to an HTTP request.
-        '''
-        if payload is None:
-            payload = prepare_payload(locals())
-        return self.session.post(f'{self.api_url}/get_company_details',
-                                 json=payload,
-                                 headers=headers)
-
 
 # Batch requests